Soft-Thinking+noise+loss-main/requirements.st.txt
accelerate==1.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
airportsdata==20250706
annotated-types==0.7.0
anthropic==0.59.0
antlr4-python3-runtime==4.13.2
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
compressed-tensors==0.10.2
cuda-bindings==12.9.0
cuda-python==12.9.0
datasets==3.6.0
decorator==5.2.1
decord==0.6.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
einops==0.8.1
executing==2.2.0
fastapi==0.116.1
filelock==3.18.0
flash_attn==2.8.2
frozenlist==1.7.0
fsspec==2025.3.0
h11==0.16.0
hf_transfer==0.1.9
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.5
idna==3.10
importlib_metadata==8.7.0
interegular==0.3.3
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jsonlines==4.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
lark==1.2.2
latex2sympy2_extended==1.10.2
litellm==1.74.8
llguidance==0.7.30
MarkupSafe==3.0.2
math-verify==0.8.0
matplotlib-inline==0.1.7
modelscope==1.28.0
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
nanobind==2.8.0
nest-asyncio==1.6.0
networkx==3.5
ninja==1.11.1.4
numpy==2.3.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.97.1
orjson==3.11.0
outlines==0.1.11
outlines_core==0.1.26
packaging==25.0
pandas==2.3.1
parso==0.8.4
partial-json-parser==0.2.1.1.post6
pexpect==4.9.0
pillow==11.3.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pycountry==24.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pynvml==12.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rpds-py==0.26.0
safetensors==0.5.3
sentencepiece==0.2.0
setproctitle==1.3.6
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
stack-data==0.6.3
starlette==0.47.2
sympy==1.13.1
tiktoken==0.9.0
tokenizers==0.21.2
torch_memory_saver==0.0.8
torchao==0.12.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.1
typing_extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
wcwidth==0.2.13
xgrammar==0.1.17
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
